graph: Move store traits to a separate file
